fix some warnings
Former-commit-id: be4263682d56b3852a9ad60b9f1d1096dd74fde6 [formerly ac1f70a733e933e3a349307e9b9451a8a3733381] [formerly 2a46ff175e77abe7263c6ce1b08740ac7154f110 [formerly c0f716d7c83f07467393a864938c41a4b83c01bf]]
Former-commit-id: 326ce40da04b887414a63c0383eb74bba6e31892 [formerly cc1979fe40d392bde305433384d06500684427f9]
Former-commit-id: 6b03c8bb25af7142617a0647a44e4a7bef099c5d

+++ Frame/SDK/Core/AFDataTable.h
@@ -114,7 +114,7 @@ public:
     int FindString(size_t col, const char* key, size_t begin_row = 0);
     int FindObject(size_t col, const AFGUID& key, size_t begin_row = 0);
 
-    bool QueryRow(const int row, AFIDataList& varList);
+    bool QueryRow(const size_t row, AFIDataList& varList);
 
 protected:
     void ReleaseRow(RowData* row_data, size_t col_num);
